refactor(design-system): consolidate shared Checkbox + SegmentedControl into design-system
The shared `Checkbox` and `SegmentedControl` duplicated richer design-system
components. Promote the design-system versions as canonical and delete the
shared ones.

- Checkbox: the design-system version has a real `<input>`, indeterminate
  state, controlled/uncontrolled modes, and form name/value. Its sizes map
  exactly onto the old shared variants (desktop = `md`: w-4 h-4 text-xs;
  mobile = `lg`: w-6 h-6 text-sm), so this is a clean visual swap.
  Migrated 10 call sites (`variant` → `size`).
- SegmentedControl: the design-system version adds roving-tabindex keyboard
  nav and proper `role="radiogroup"`/`role="radio"` + `aria-checked` semantics.
  Migrated the one call site (EngineSelector); `ariaLabel` → `aria-label`.
- Deleted both shared components; updated the shared index and README. The
  unrelated `segmentedControlClasses` utility is untouched.
- Test mocks that stubbed the shared modules now partially mock
  `@/design-system`; EngineSelector assertions moved from `aria-pressed` to
  the design-system's `aria-checked`.
