+++ src/shell/Mobile/MobileSettingsPanel/MobileSettingsPanel.tsx
@@ -9,13 +9,12 @@ import { ConfirmDialog } from '@/shared/components/ConfirmDialog';
 import { HalfGridModeBlockedModal } from '@/shell/Modals';
 import { DeferredNumberInput } from '@/shared/components/DeferredNumberInput';
 import { PrintBedInput } from '@/shared/components/PrintBedInput';
-import { Checkbox } from '@/shared/components/Checkbox';
 import { SectionHeader } from '@/shared/components/SectionHeader';
 import { SettingsRow } from '@/shared/components/SettingsRow';
 import { LoadingFallback } from '@/shared/components/LoadingFallback';
 import { lazyWithRetry, namedExport } from '@/shared/utils/lazyWithRetry';
 import { useTranslation } from '@/i18n';
-import { Button, Stepper } from '@/design-system';
+import { Button, Checkbox, Stepper } from '@/design-system';
 import { UserDock } from '@/shared/components/UserDock';
 import { useFeatureFlag } from '@/shared/hooks/useFeatureFlag';
 import type { SettingsTabId } from '@/shell/Modals/SettingsModal/types';
@@ -164,7 +163,7 @@ export function MobileSettingsPanel() {
               {t('mobile.settings.allow05UnitPrecision')}
             </p>
           </div>
-          <Checkbox checked={halfGridMode} variant="mobile" />
+          <Checkbox checked={halfGridMode} size="lg" />
         </div>
       </section>
